zfs_rename: support RENAME_* flags
Implement support for Linux's RENAME_* flags (for renameat2). Aside from
being quite useful for userspace (providing race-free ways to exchange
paths and implement mv --no-clobber), they are used by overlayfs and are
thus required in order to use overlayfs-on-ZFS.

In order for us to represent the new renameat2(2) flags in the ZIL, we
emulate the renameat2(2) operations using several transaction entries.
This allows us to avoid adding new ZIL types for something which is
Linux-specific -- and the operations are safe because of the context in
which a ZIL replay will be executed. RENAME_EXCHANGE requires a bit more
work because we need to create a temporary file with a random name. See
the comment in module/zfs/zfs_log.c for more details.

This feature was worked on by both Aleksa and Pavel. Aleksa did the
initial implementation, Pavel reworked the ZIL handling to not require
new ZIL types and various other improvements, and then Aleksa did the
final rebase and carry.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>
Signed-off-by: Pavel Snajdr <snajpa@snajpa.net>
Co-authored-by: Pavel Snajdr <snajpa@snajpa.net>
Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>