renameat2: add EL7 compatibility handling
EL7 has .rename2 op upper in the inode_operations_wrapper, so let's add
a wrapping ifdef to dir inode_operations and ifdef the ops assignment
too.

Configure test lives in kernel-rename.m4 for now, as rename2/EL7 case is
the only reason why we need it - for now, until support for tmpfile is
needed. then it would me more logical to move this check to a separate
file.

Signed-off-by: Pavel Snajdr <snajpa@snajpa.net>
Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>