dcache: remove zpl_revalidate entirely
Open files, which aren't present in the snapshot, which is being
roll-backed to, need to disappear from the visible VFS image of the
dataset. Kernel provides __d_drop function to locklessly drop invalid
entry from the dcache, but inode can be referenced by multiple dentries.
The introduced spl_d_drop_aliases function walks and invalidates all
aliases of an inode.

This is necessary for overlayfs support, because overlayfs doesn't
support lowerdirs that have a d_revalidate callback defined (which is
traditionally only remote filesystems).

Signed-off-by: Pavel Snajdr <snajpa@snajpa.net>
[Aleksa: rebase and remove old zpl_revalidate code.]
Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>